Adjust fix-me tests
